doc: Automatically determine the number of examples in the gallery page (#3301)
In the GMT gallery page (`source/gallery.rst`), we include all examples using a for loop:
```
{% for i in range(1, 51) %}
```
The total number of examples is hard-coded here. We have to increase the number
when new examples are added.

In this PR, the total number of examples are automatically determined by counting the number
of `ex*.rst` files in the `doc/rst/source/gallery` directory. Thus, we no longer
need to manually increase the number.