Deeper check that an argument is a constant (#3321)
The new DAYNIGHT operator expects the given lon,lat to be constants while the width can be variable.  However, the way grdmath keeps track of things is not ideal.  A given constant is indeed flagged as a constant, but if we do some mathm like subtracting two constants, the result may be stored in a grid and the constant bool is now false.  That does not mean the values are not constant, of course.  This PR adds an extra check to detect these cases before DAYNIGHT is allowed to run.