doc: Fail the CI job if sphinx reports any warnings (#3337)
Currently, sphinx doesn't report any errors and warnings (e.g., wrong reST syntax,
include a non-existent file, missing references) when building the GMT documentation.
That means our documentation is in good shape and we should keep it warnings free.

When we make a non-fatal mistake in the documentation, sphinx will report a
warning, but the exit code is still 0. Then the CI job will think the sphinx build is
sucessful, and we won't be aware of the mistake unless we check the CI report
or build the documentation locally.

This PR fixes the issue by checking if the warning log file "html.log" is empty.
A non-empty html.log file means sphinx reports some warnings, then the CI job fails.