pssac: Fix a bug when reading a SAC data in a partial time window
When reading a SAC file in a time window from t1 to t2, the timings of
the data points may not exactly match t1 or t2, thus we must be
very careful with the begin time of the new data segment.

The old code always uses t1 as the begin time of the new data, which may
cause a tiny time error < dt (dt is the time interval of the data).

The error is usually tiny and ignored, but for studies that require
high-precision data timings, it's very important to keep the data
timings.

Also fix a failing test due to the tiny time shift.