Ensure batch_init.sh has the number of jobs (#3351)
If -T is given a file that is actually created by the preflight script then we dont know the number of jobs when the batch_init.sh settings are written.  This does not matter in the similar case for movie, but here, the postflight script is run after all jobs are completed and it may really need to know.  So if it si unknown we simply rewrite it with the known job count after the preflight scirpt completes.