Help user avoid common movie/batch mistakes (#3343)
* Help user avoid common movie/batch mistakes

All teh MOVIE_* or BATCH_* script variables must have a leading $ or % to actually give a value.  It is pretty simple to forget to do so, and then you just have a random text in your script.  Now, both batch and movie looks for that. We also recogonize the deprecated ways to do a sub-shell (`,...`) and recommend $(...) instead.
Finally, the common functions used by both batch and move were incorrectly named gmtlib_*.  They need to be gmt_* per our convention.

* Also check for unbalanced {}

* Bad merge

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>