grdinfo -C now also includes registration
For whatever reason, wihle grdinfo t.grd reports a line about the node registration, there was no such item when we asked for a numerical record -C.  So no way of easily determining if a grid is pixel or gridline in a script, other than grepping on the text output.  Because the order of outputs were set a long time ago, this PR just adds registration as the very last item of the record.