Restore initial offsets in PSL (#3367)
See #3365 for background.  The problem was that psxy plots decorated lines by drawing the line first, then call it self to plot the symbols.  In the second call we do a straight overlay so the x/y offsets passed to gmt_plotinit (and into PSL) are zero.  Yet, when we return back to the original psxy call, we want to undo any translated done by that command.  Alas, those values were wiped when 0,0 was given.  The solution is to let PSL remember the origin value before the update, and then restore the values when we call gmt_plotend.  Fixes this issue and all tests pass.