Update Fedora install instructions
Fedora is a continously-evolving distro and has a very short release cycle.
Fedora usually releases a new version every ~6 months, and maintains each version for ~13 months.

Currently, the supported Fedora versions are:

- Fedora 30
- Fedora 31
- Fedora 32
- Fedora Rawhide (testing/unstable)

Fedora >=31 are providing GMT 6.0.0, and are actively maintained, and
Fedora 30 will EOL at the end of this month.

Here we recommend Fedora users to use the official GMT packages.