Add a github action to automatically make a draft release after tagging
This PR adds a github action to automatically make a draft release for us.

Making a tag (e.g., 6.1.0) and pushing it to github will trigger the
github action. It will:

- Download 5 packages (source tarballs, bundle and windows installers) from the GMT FTP
- Create the checksum file
- Make a draft release
- Attach the 5 packages as assets of the draft release

The whole process only takes ~2.5 minutes and is fully automatic.
We still need to review the draft release and publish it if everything
looks good.

**NOTE**: Since the github action needs to download packages from the GMT FTP.
We need to upload the packages before making a tag.

The release checklist is also updated accordingly.