Extend -D in subplot fix to all -D modules
The problem solved in #3454 is actually shared among found modules that all have the same -Dx option: gmtlogo, psimage, pslegend, and psscale.  The #3458 issue made that clear to me.  So I just generalized the fixed to do the same for any of these four modules in the same situation (-Dx inside subplot panel) and the problem was solved.  Closes #3458.