Fix inconsistencies in new -F option
The -F option scans all cross-profiles made via -C and report on some aspects of each profile as a function of the along-track (not cross) distance.  THus, the first output column needs to be the along-track distance.  Furthermore, the initial implementation skipped output records if no values could be determined. Now we follow standard GMT prqctice of issuing NaN values if they cannot be determined.