Let symbols with error bars be shown if bar are inside window (#3471)
* Let symbols with error bars be shown if bar are inside window

This addresses the issue raised on the forum.  I have fixed this so no need to use -N to see symbols insude the box.  However, the users y-errors goes negative which cannot happen for log10 and taht is why this looked odd.  I now draw the y-bar only down to the point if the lower value is NaN.

* Update ex03.ps