Issue correct message regarding w/e/s/n adjustment (#3494)
See #3488.  Turned out this was not actually bad rouding.  The comments say we extend the grid outwards adn we do.  however, the arning message out resetting e to 12 is wrong.  It reports the result before we have made the adjustment.  I have fixed this so we report the actual region used.  I am unable to reproduce the problem the original poster reported and I assume it was fixed by an earlier PR.