PostScript colormasking affecting non-PS output via -A (#3495)
PostScript has its own way of dealing with transparency: Find a unique unused color and set all pixels that should be transparent to that one.  Fine, but that is now how rasters work, so for -A none of those things should be triggered.  The end result looked fine: a transparent PNG.  However, while transparency was 255 for the right pixels, their colors were actually red, and when Google Earth resamples them the red leaks a bit.  This could be a bug in GDAL or Google Earth but surely we should not set those pixels red in the first place.