Fix incorrect use of gmt_M_grd_x_to_col/gmt_M_grd_y_to_row
These return the nearest node in a pxiel or gridline grid.  However, when used to convert a node coordinate to a node column or row, you must actually pass the correct coordinates.  We had sevearl places were we passed west +k * inc but that is only a node coordiante for gridline grids.