Prevent pen settings for fronts to leak back to current pen (#3533)
See #3531 for context.  The probelm was that the gmt_plot_front could change the pen thickness to darw the outline of a symbol, but then this would leak back to the main module and the next line-segment would inherit that thickness instead of what -W had said.  The solution, as for decofated lines, is to surround the front plitting with a matching gsave/grestore commands.