Fix use and doc on invisible lines
See #3518 for context.  First, internally we had renaemd that modifier +i but we do accept +h for backwards compatibility.  +h is generally used for header or other things so +i is better. However, this did not (1) make it into the documentation, (2) the usage message, or (3) actual code [which had a bug].  Closes #3518.