Ensure movies made with FFmpeg receive even raster width (#3542)
If users supply their own custom -C settings it is possible they end up with a width or height translating to an odd number of pixels.  This does not work wtih FFMpeg.  Now, instead of giving an error message but continuing on, we simply nudge the width and/or height upwards in small steps until the nearest pixel is one larger, and give a warning instead. Closes #3540.