Eliminate misleading pointer from blend structure
The HH pointer in the blend array was set to point to the hidden structure in the grid header and used once.  But then we destroy the grid and now the pointer is pointing to junk.  Could be confusing down the road.  So now removed and we use a single variable to get the hidden content in the only place where it is needed.