Let the auto-determined region be consistent
When no -R -I is given we determine the max extent wesn, and if all grids have the same inc we can create the output grid region and use that inc.  However, we had no defense against these different wesn being phase-shifted.  Now, we ensure that the region we select has an integer number of steps in dx between min and max for both dimensions.