Only make tiny adjustments to dx and dy if both need it
We have a roundoff-patrol checker that looks to see if an geograic grid increment is close to a multiple of integer seconds or minutes.  if it is close enough then we set it to the exact.  however, the case that triggered this discussion was close enough to 3s in x but not in y.  I think we should at least require both increments to be adjusted or neighter, and possibly tighten the decision when to do this.