Correct ex47 rows
Ex47 derived from an earlier test (regress_2.sh) which was written in classic mode.  The translation to modern mode mixed up which rows we are doing, so the initial order (bottom to top in classic) was roxy while for the modern mode going down the rows it needs to be the reverse: yxor.  It was roxy except for the L1 column which was yrox.  So bug wiwthin a bug.