Properly set parameters for cyclic CPT symbol that scale with bar size (#3665)
* Properly set parameters for cyclic CPT symbol that scale

The derived parameters of the CPT cyclic symbol did not scale correctly with bar thickness and only even worked for moderatelly thin bars such as the test examples in psscale.  A wider bar and we either got ugly symbols or a crash.  Also, the symbol was always black and invisible if we wanted white on black.  This PR accomplishes both of these:

We use the color of MAP_FRAME_PEN to determine the color to use
We define the length and width of the arrow head as fixed fractions of the circumference (and hence the radius or bar thickness).

* Update PS with cycle symbols

* Update psscale.c