Add new bitflag GMT_SESSION_NOHISTORY to turn off shorthand and history (#3671)
* Add new bitflag GMT_SESSION_NOHISTORY to turn off shorthand and history

Developers may wish to not rely on the file-based communication between modules via the gmt.history file and the -R -J shorthands.  TO do that they need to pass the new GMT_SESSION_NOHISTORY bit-flag to GMT_Create_Sessions.  Note: gmt.c does not pass this value since we rely on gmt.history.

* Update gmt_enum_dict.h