Let grdcontour -N still be a oneliner (#3702)
THe problem is that grdcontour -N calls grdview and then itself via GMT_Call_Module.  However, we have already flagged it as a oneliner but now there are several calls.  The solution was to not trigger the oneliner check for grdcontour -N and then detect it instead in the special -N branch in grdcontour.  We parse the options and discover if the command was a oneliner or not, and if it is we craft the gmt begin call needed and add the gnt end show call at the end.  Seems to work.