Address bug in grdseamount -M option
The writing of the filenames etc to the list file with -M crashed because we did not pass theGMT_WITH_STRINGS when creating the dataset.  I also added time as the first column since the initial version of this went back to text output days (it was all a string record) but it makes sense to write the time as a floating point, followed by the filename and the formatted time label.