gmt_make_enum_dicts.sh needed some TLC
It had an agreesive grep -v to avoid lines that start with struct etc but it would also trigger if the word struct appeard in a comment.  Added ^ to indicate start of line.  This picks up the missing ones.  GMT_GRID_ALL is a backwards compat for < 5.3.3 so wont be included (it is not an enum, just a #define).