Only set geographic column formatting if in geographic mode
See issue #3695 for details.  The problem was that -N was not consulted when setting teh output column types.  In Cartesian mode they are all regular floats, while in geographic mode we need to set special formatting for longitude and latitude.  This PR turns off these column operations when -N is set.  Closes #3695.