Make sure all test scripts have a set GMT_SESSION_NAME
While required for modern mode script, the fact is that any GMT command starts its day by checking if it is run within a modern mode environment.  It learns taht it is by finding the special temporary directory created by gmt begin.   Hence, if ctest runs tons of tests and the classic ones do not have GMT_SESSION_NAME and default to $$, we may invertantly ahve to sripts that share th same session name.  This seems to happen because some tests fail when run in parallel by claiming they are modern when in fact they are classic.  This PR sets the GMT_SESSION_NAME for any script run as a precaution, and so far it seems to have fixed my problems (running 24 tests in parallel).