Let libcurl downloads be managed via lockfiles (#3735)
* Try to use filelock for downloads

We want to avoid race conditions on gmt_data_server.txt and gmt_hash_server.txt.

* Update gmt_remote.c

* Delete a failing gmt_data_server.txt file so it can be regenerated

If for whatever reason the gmt_data_server.txt ends up blank, reading will fail.  Rather than just complain about it, we now delete the file since we (a) know it is broken and (b) unless it is removed we will not attempt to refresh it for another cycle (by default 24 hours).