Do not let the API object for the assembled grid live one
GMT_Read_Data assigns a new object number to what is read, but for tiles it calls gmt_assemble_grid which also creates an object to pass to grdblend.  However, once we get it back from grdblend we only want to keep the grid and discard the container since GMT_Read_Data already has a container for it.