[Backport 6.1] File lock on CURL downloads (#3808)
* Let libcurl downloads be managed via lockfiles (#3735)

* Try to use filelock for downloads

We want to avoid race conditions on gmt_data_server.txt and gmt_hash_server.txt.

* Update gmt_remote.c

* Delete a failing gmt_data_server.txt file so it can be regenerated

If for whatever reason the gmt_data_server.txt ends up blank, reading will fail.  Rather than just complain about it, we now delete the file since we (a) know it is broken and (b) unless it is removed we will not attempt to refresh it for another cycle (by default 24 hours).

* Exempt URL queries from having lock files (#3768)

Since the URL of a quiery is not a good file name, we do not create an advisory lock file for such downloads.  Addresses #3765 hopefully.

* Checked wrong string for URL (#3780)

Hopefully fixes #3765.

* Forgot the other place where locking occurred (#3785)

Now both places where lockfiles are used avoids URL queries.  Closes #3765.

* Must close file before delete it (#3804)

Co-authored-by: Paul Wessel <pwessel@hawaii.edu>
Co-authored-by: Joaquim <jmfluis@gmail.com>