Handle matrix vai reference for grid input
Testing is done via testapi_matrix_360_ref.c  Seems to work but raises an issue related to memory freeing.  Since the matrix data is only pointed to by the grid, I set the grid alloc_mode to external.  However, that also means that the x and y arrays that are created when creating the grid is left unfreed at the end.  We could just ignore this rare memory leak or do something about it.  One way is to add info to the struct HIDDEN so that we free those arrays even those the matrix is left untouched.