Only turn on clobber background if it can work (#3839)
* Only turn on clobber background if it can work

See #3802 for background.  The problem was that pscoast decided that if both ocean and land will be painted, we can save some work by painting the entire canvas with ocean color since then the lakes are open holes that see through to that color.  However, if -C is sued to change the color of lakes then of course we must paint the lakes individually anyway.  The code did not check for this csae.  I added a new function gmt_same_fill that carefully determines if two fills are the same or now, and if lake fill and ocean fill are different then we cannot use the simplified clobbering.  Closes #3802.

* Update pscoast.c