Updating a few keyword-value pairs for common options (#3842)
* Updateing a few keyword-value pairs for common options

I added the relatively new -qi (--read-rows) and -qo (write-rows)  and -c (--panel) to the mix,  Also updated the information for --lengend since we made some major changes since that was first coded.  Tested these informallly to make sure they worked in the translation from long to short syntax.  Also made a constant for the row used for grid increments (-I) since it is a near-global option that will be needed by many modules, so to avoid typos we will use GMT_INCREMENT_KW.

* Update gmt_constants.h