Better recognition of pi-number (#3941)
* Better recognition of pi-number

We had a very lame check to determine if a token was a pi-number which simply checked for the string "pi" in the token.  That worked fine until we hit the Greenland town of Aappilattorq.  Here the pi was found it we obtained the nubmer inf...
We now do a full check to determine if it really is a pi-number via new helper function gmtio_is_pi.

* Update gmt_io.c