Let z-axis annotations be normal as an option
The y-axis by default plots annotations normal to the axis, but accepts +ap to plot parallel to axis.  The z-axis only plotted parallel to the axis and did not accept +a.  This PR allows +a but keeps the default +ap for backwards compatibility.  Use +an to change it, or use  --MAP_ANNOT_ORTHO=wez.