Draw both x- and y-gridlines on the 3-D xz and yz back-walls
PReviously we only drew the horizontal z-gridlines on the back-wall.  There was a good reason for this.  Given how our 3-D perspective is done (via clever matrix rotation of the view point) we cannot draw the constant x and y lines on the backwalls for geographic projections.  However, we can do it for Cartesian projections.  So this PR draws what we can do, thus adding a full grid on Cartesian backwalls.