Adapt to new format of Ghostscript registry key.
Ghostscript now writes a registry ey with the patch number, E.g. 9.53.1 while before there only up to minor (9.53). This broke the ability to find that key in the Windows registry. This PR fixes it but at the expenses of not searching anymore to possible multiple GS versions installed. Now it takes the first one it picks.