Tiny changes on commenting for PostScript text, plus doc update (#4250)
* Marginally improve comments

The PostScript text typesetting is black magic and too hard to penetrate for me now.  I added some more commets but this is not easy to understand, being all in reverse Polish notation to boot.  For now we will simply document that the paragraph mode does not support composite characters.

* Update text_common.rst_