subplot set fails if not given index|row,col
Per the subplot documentation, gmt subplot set without any row,col or index shall go to the next logical panel, but in reality it crashes by trying to read the next option.  This PR checks if the next option is a non-option argument (hence flagged as "input file") and only then tries to parse.  Otherwise we default to the next panel.