Plot bar-symbol errorbars on top of symbol
Unlike regular (smaller) symbols with error bars, a bar symbol usually goes all the way from the axis to the value.  Plotting an error bar behind it means the lower bound of the bar will be obscured.  Here we plot the errorbar on top if bar symbols are used.  This is now clarified in the docs.