N/S ticks for "global" conic maps requires a special check
For conic map where w/e is 360 we do not end up with a circle shape but a pacman.  Hence, computing the angle of the ticks on the east and west boundary cannot use the general algorithm since it will jump across the pacman gap.  Instead, we use our knowledge of the situation to fill out the tick locations and the angles directly.  Fixes a bug and updates two PS plots