Improve CMake scripts if sphinx-build is not found
sphinx-build is required for building documentations. If sphinx-build is
not found, there is no need to make the verbatims and convert PS to PNG
formats.

In the old CMake scripts, if sphinx-build is not found, cmake gives the
following error:
```
ninja: error: 'doc/examples/_docs_rst_mkdir_verbatim', needed by 'doc/rst/source/_verbatim/anim01.txt', missing and no known rule to make it
```
The error message means that CMake wants to create the verbatim
`anim01.txt`, but the directory `doc/rst/source/_verbatim` (i.e., the
target `doc/examples/_docs_rst_mkdir_verbatim`) is not found. The error
message is very confusing.

This RP should fix the issue. If sphinx-build is not found, it should
give the warning like `Target docs_html is not defined.`