Several problems in grdgradient (#4328)
* Incorrect header used to compute nodes if external

grdgradient computes derivatives and requires at least a one row/col padding for the algorithm to work: We do the finite differencing on a bin with 4 nodes and store the result in the top left node which is not used in further calcuations.  At the end we restore the padding so the output is correctly aligned.  For files (automatically having a pad of 2) this works fine.

* Update grdgradient.c

* Update grdgradient.c

* Update grdgradient.c