Better handling of error conditions due to map_setup call
We had a convoluted set of error checks and reporting that gave both correct error messages and confusing wrong references to other things due to a mix of internal and "official" error codes.  This PR changes this to a simpler reporting.  I was unable to reproduce the SEGV in the OP examples but I believe it was related to bad array indices into the gmt_error_string[] array.  This hopefully goes away - if not we will follow up.