Fix duplicate array values for most modules (#4370)
* Fix duplicate array values for most modules

If an array is given via a comma-separated list or a file, for most modules that acquire such an array it is required that it be monotonically increasing and not have duplicated values.  This we never checked for. So most of the PR involves adding a new flag that forces sorting and removal of duplicates and then passing that flag to most places where gmt_parse_array is called.  The only exception I could find is gmtmath which is fine with a list of numbers.  A minor fix in histogram had to do with updating the wesn for x after parsing the list.

* Update gmt_support.c