Link docs to the master branch
Currently, the "Edit On GitHub" button in the documentation is linked to different
branches for different documentation versions. For example, the v6.1 and
dev documentation is linked to the 6.1 and master branches, respectively.

It doesn't work as we expect.

1. We no longer make changes to the 6.1 branch. However, the default
   branch is still 6.1 branch if users make edits to the documentation
   by clicking the "Edit On GitHub" button (see #4402).
2. We decided that, every PR should be merged into master branch, and
   then backported to 6.x branches if necessary.

Thus, the "Edit On GitHub" button should always link to the master
branch.