Enforce --n+a in grdimage if given categorical CPT
Hardly any GMT user will know apriori that when they image a geographic categorical grid that they must use -nn+a to avoid interpolation.  This PR fixes that by warning users if those settings are not selected and overriding them with what is required.  The documentation has also been updated to discuss this and other options.,