Enforce --nn+a in grdimage if given a categorical CPT (#4427)
* Enforce --n+a in grdimage if given categorical CPT

Hardly any GMT user will know apriori that when they image a geographic categorical grid that they must use -nn+a to avoid interpolation.  This PR fixes that by warning users if those settings are not selected and overriding them with what is required.  The documentation has also been updated to discuss this and other options.,

* Update grdimage.c

* Update grdimage_notes.rst_

* Update src/grdimage.c

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>