Use named constants rather than 1 and 2 to flag categorical CPTs
It is hard to read code like P->categorical & 2 unless you know what it means.  This PR just introduces  the named constants GMT_CPT_CATEGORICAL_VAL and GMT_CPT_CATEGORICAL_KEY for these two numbers.  We also add a check to not write back and foreground colors to a CPT when the CPT is categorical or cyclical.