Avoid adding a step for non-integer seconds
To avoid roundoff we always added an increment to the flaoting point seconds to make sure gmt_gcal_from_dt would not suffer a round-off by having time like dt = 123234234234.999999999 be truncated. Adding 0.5 ensures we get the right second when we floor.  However, with time with precisions higher than whole seconds, say dt = 123123123.75, we do not want to add 0.5 since the value is already clearly between floor and ceiling of the right second.