Settle on transition size from old to new psl_shorten_path algorithm (#4478)
See old issue #439 for background.  Playing around some more I found that using a cutoff of 50,000 points in a polygon affects none of our tests but it does fix the original problem (which is a very large polygon).  I think this is a reasonable compromise for now.  I added a reference to this issue in the postscriptlight.c code where this is addressed. Closes #439.