Address failure of psxy[z] to deal with -Z data (#4485)
See #4482 for background.  The problem was that we needed to read a single-column file (hence x-column) as z-data, but we failed to temporarily change the column type from, say, longitude to float.  So for projections, all the data got a modulus 360 treatment and possibly a 180 sign shift...
I also took the oppotunity to give more meaningful names to functions get or set the column type.  Instead of gmt_set_column it is now gmt_set_column_type, which makes it easier to find.  Also added gmt_get_column_type and used it in this PR.