Prevent ESRI checking from accepting a .jp2 tile (#4492)
* Prevent ESRI checking from accepting a .jp2 tile

THe checks in the gmtlib_is_esri_grid was defeated because the file name ahd some similarities with the speecial case for *.hgt files in there.  Now we refuse to even look inside *.jp2 files and punt back to GDAL.

* Make gmtlib_make_equidistant_array more robust

For the csae of one layer in grdinterpolate I am guessing the inc may be zero, leading to a huge n that swamps the memory request.  This is now caught to set n = 1.

* Update gmt_support.c

* Initialize inc[GMT_Z] to zero

* Init wesn and inc