Grid and cube containers allocated to hold matrices should not allocate data
When grids or cubes are passed via external matrices we create a GMT grid or cube container to hold the information.  However, whether the data array should be allocated depends on whether it is a reference or duplication. Hence, all the GMT_Create_Data calls for these situations should pass GMT_CONTAINER_ONLY so that any allocation later is allowed.  Hopefully closes #4577.