Update -D to include z-dimension, optionally set var name
In the process, change greenspline -D to -Z but backwards compatible so that it and grdinterpolate can take -Dinformation to set the various grid/cube variables.  Updated scripts using -D and ensure the "name [unit]" parsing works for x,y,z, and d.  Renamed +zsunit to +ddataunit for both grids and cube but this is backwards compatible for grids of course.