Enable writing of 3-D netCDF data cubes from greenspline and grdinterpolate (#4581)
* Let gmtapi_export_cube call gmt_write_nc_cube

Thus, the work on writing a 3-D cube calls to gmt_write_nc_cube in gmt_nc.c

* Properly name the cube functions

* add cube arg to gmtnc_grd_info so we can deal with 3-D vars

* Add netcdf var ID for 3-D time/depth variable

* Fix wrong number of output levels in grdinterpolate

* Set cube data type

* Getting close

* Update greenspline to write cube command

* Remember netcdf IDs after reading

* Got a working solution but z listed before x,y in dimensions list

* Use UNLIMITED dimension for layers

* Fix KEY for greenspline to be U if cube output

* Remove initial STACK option to read grids directly into cube

This does not work since registration is a problem.  Now done in grdinterpolate instead.

* Document the reformatting of stacked grids to cube

* Update grdinterpolate.c

* Update gspline_4.sh

* Improve the docs

* var changes

* Fix nc parameter IDs

* Add gmt_nc_is_cube and fix geometry mistakes

* Add gmt_nc_is_cube function

* Let grdinfo -Q handle cubes

* Fix z subset glitch

* Update gmt_api.c

* Fix reading cubes with selection of variable

* fix KEYS for grdinfo

* Towards resolving data type info

* Fix unit issues for cubes

* fix chunking array

* Add tests for 3-D cube writing

* Update -D to include z-dimension, optionally set var name

In the process, change greenspline -D to -Z but backwards compatible so that it and grdinterpolate can take -Dinformation to set the various grid/cube variables.  Updated scripts using -D and ensure the "name [unit]" parsing works for x,y,z, and d.  Renamed +zsunit to +ddataunit for both grids and cube but this is backwards compatible for grids of course.