MInor under-the-hood changes related to long-options
I fixed the issue renaming read-cols to incols and write-cols to in-cols as these are a bit specia,l since both are -q in short-form but map to -qi and -qo.  Also, the description of long directive and argument was a bit off in the gmt_common_longoptions.h file,  I also renamed gmtinit_kw_replace to gmtinit_translate_to_short_options and added a dummy inverse function gmtinit_translate_to_long_options we will use to test the translation tables.