Let grdcontour -A+T give compativility warning and not an error (#4671)
* Let grdcontour -A+T give compativility warning and not error

In GMT 6 we eliminated the +T modifier and only use +t since it seemed pointless to have both as users can just sue teh columns they need.  However, it was a mistake to let +T trigger an error sicne it used to be a valid modifier.  This PR allows +T under compatibility with version 6.

* Update gmt_support.c