Make it clear that placing -Sq in seg headers does not eliminate need for -S
See forum post for context. In review it was not clear that -Sq is still needed on the command line to tell psxy we are plotting a symbol; just having these in the segment header is too late as now we are plotting lines...
I updated the docs to make this clear.