Add developer doc chapter on refactoring (#4674)
* Add section on our refactoring plans

The "C refactoring" work proposed and funded by NSF was never well described iin details, in which the devil lies of course.  Here, I start the documentation of why refactoring was proposed, what its goals are, and discuss the various traps we need to avoid in designing how the next level API will look like and be useful across the GMT ecosystem.  It is expected to be updated until we reach consensus.

* Update refactoring.rst