Reset azimuth for degenerate ellipses and rotated rectangles (#4683)
* Reset azimuth to zero for degenrate cases

When -SE- and -SJ- is used to set a degenrate ellipse (circle) or rectangle (square) we are only given the single dimension and duplicate it for both dimensions.  However, the argument remained in the azimuth position which was never reset to zero,  While not causing problems, it is ocnfusing to debug cases and find an azimuth of 2000, for instance.  This PR fixes this minor issue.

* Update events.ps