Correct Cartesian ellipse output in project and allow degenerate case
The -Z option in project can be used to create the coordinates of an ellipse in either geographic or Cartesian space. Unfortunately, the Cartesian case had a bug in that it assumed we gave it semi-axes, not full axes like for the geographic case as well as the case in psxy.  This is now fixed.  In addition, I made it simpler to specify circular output (the degenerate ellipse) but just providing the diameter (again, like in psxy -SE-).