Control media cropping via PS_CONVERT settings (#4735)
* WIP Control media cropping via PS_CONVERT settings

See #4734 for context.  This PR consider the session and figure settings for the psconvert options and if containing A+n (no cropping) we replace that with A+M when passed to psconvert.  This undocumented option leads to cropping to the current default media size (set by PS_MEDIA).

* Update gmt_init.c

Let any pre-session PS_MEDIA setting be ignored.

* Handle initial external gmt.conf files and PS_MEDIA

* Update gmt_init.c

* Use constants for both margin settings

* Update gmt_init.c

* Update gmt_plot.c