Fix movie shift when back/foreground is used and -Xc -Yc is set (#4736)
* Fix movie shift when back/foreground is used and -Xc -Yc is set

See #4733 for background.  The issues was that while GMT normally adds layers upon layers to build a final PS file, in the special case of movie and -S we are in fact building complete plots (the background and foreground plots are complete PostScript files that can be viewed, say, during debug.  However, before of this they have headers which must be skipped when psconvert assembles the final plot. Furthermore (and this part was missing) the memory of what the margin location is (PSL_x_orig, PSL_y_orig) needs to be reset to zero.  With that fix we are in business.

* Update anim05.sh