If -f indicates abstime then dont fuss that -J does not (#4749)
* If -f indicates abstime then dont fuss that -J does not

We used to get annoying messages from gmt_mapsetup if -f had indicated absolute time but -JX did not have that 'T'.  it was a warning, but then it continued, taking no corrective action.  This PR now does two things:
Sets the missing linear projection type to GMT_TIME
Fixes an issue in the auto-interval routine so that it knows the axis is time,

* Update gmt_io.c