psxy drew ellipse outlines when not requested
See #4750 for background.  In psxy, we may have to plot an ellipse twice if it is near a periodic boundary.  in those cases the fill and outline must be placed separately.  however, there was no check that the outline had been requested, leading to the bug.  3 of the 4 tests that used these ellipses were black so not possible to tell.  Closes #4750.