Fix bug in PSL library related to writing HSV colors
The PS_COLOR_MODEL settings [rgb] can be set to hsv and then color specifications written in the PostScript uses the sethsbcolor operator instead of setrgbcolor.  Both requres triplets in 0-1 range but for hsv we kept hue in angle 0-360, resulting in the wrong color.  This PR fixes this issue but the discussion will continue (#4753).