Fix horizontal line-angle in contouring (#4782)
The computed line-angle for a symmetric case (e.g., top of a parabola) where the answer should be horizontal (0) we had a dumb special check that said 90.  atan2 is perfectly capable of doing this for us.  This fixes an unnoticed error in gspline_3.ps where one annotation was 90 degree rotated.