Ensure that single frame movie or jobs get non-zero precision (#4786)
The precision is used to determine the width for the integer tag for each job, e.g. 001 vs 01.  However, if the user's choices resulted in a single job or frame, the precision calculation (log10(1)-related) gave 0.  Now I add a bit noise so the ceil will promote it to at least 1.  See #4779.