Distinguish between geographic and separate lon,lat wrapping (#4802)
We were abusing the macro gmt_M_is_geographic to determine if x is longitude but that macro also checks that y is latitude.  In many of the instances we simple wanted to use gmt_M_x_is_lon or gmt_M_y_is_lat separately, which was the trouble in #4795.  Closes #5795.