Check how close to pole we are with -JOa (#4803)
* Check how close to pole we are with -JOa

This particular formulation of the oblique Mercator takes a point and an azimuths and then computes a second point in that direction that is 10 degrees away from the first.  Well, if the first point is close to a pole then we might end up on the other side of the pole, doing some major longitude shifting.  This is what caused the problem in #4798.
This PR checks the distance to the pole and makes sure we do not use a distance larger than |90-lat|.  I added a new test to make sure this remains functioning.
 Closes #4798.

* Add PS original