Let histogram -C apply to either bin center of bin value
Up through 6.1.1 the color lookup was based on mid-point on the bar, but I changed that to the bin value instead (count, percent) in master without providing a path to the original scheme.  This PR restores the original scheme as the default and adds the other scheme via modifier +b to -C.  I have updated flip.sh test to show both in action and also fix the lack of a logarithmic CPT there.