Improve spacing of parallel gridlines
WHen the user selects automatic gridline interval (e.g., -Bag) then we determine a suitable interval separately for lon and lat, then simply used the largest.  That works well for large map regions, but if you are doing a small map near a pole, and especially for oblique maps where the pole is inside the map, the longitude range of 360 may swamp the latitude range. This leads to large parallel spacing of 15-30 degrees even if the map is just a few degrees of latitude around a pole.  This PR checks for these cases and allows for separate intervals when the situation requires it (for annotations, ticks, and gridlines).  Thankfully, only 8 PostScript files needed to be updated; they mostly were polar views with basically no parallel gridlines except the polar cap.