Finally address the need for increasing annotation offset with latitude (#4852)
* Address the need for increasing annotation offset with latitude

Projections like Hammer and WInkel has a west and east boundary that curve from Equator to the poles, being vertical only at the Equator.  As the boundary becomes more and more inclined, the horizontal latitude annotations start to encroach on the boundary.  This PR addresses this by adjusting for the extra space needed due to the sloping interface and annotation font.

* Update gmt_plot.c

* Update gmt_plot.c