Merge -C with -D and improve parsing
TO make room for a possible -Ccpt option across many modules, we let the optional alternate point be set via a modifier to -D (which sets the primary point). The new parsing allows for long/lat to be given in the usual geographic syntax accepted elsewhere in GMT (i.e., ddd:mm:ssF as well as decimal).