Let movie and batch use input script mode to decide the OS
When running the windows-installer versions of movie or batch the user may wish to write scripts in bash or batch.  Thus, commands for moving and copying files as well as directory separators should be determined based on the script type and not just the fact it is a windows binary.  Hopefully this PR will address #4821.